+++ .editorconfig
@@ -1,40 +1,40 @@
-root=true
+root = true
 
 [*]
-charset=utf-8
-trim_trailing_whitespace=true
-insert_final_newline=true
-indent_style=space
-indent_size=4
+charset = utf-8
+trim_trailing_whitespace = true
+insert_final_newline = true
+indent_style = space
+indent_size = 4
 
 # ReSharper properties
-resharper_xml_indent_size=2
-resharper_xml_max_line_length=100
-resharper_xml_tab_width=2
+resharper_xml_indent_size = 2
+resharper_xml_max_line_length = 100
+resharper_xml_tab_width = 2
 
 [*.{csproj,fsproj,sqlproj,targets,props,ts,tsx,css,json}]
-indent_style=space
-indent_size=2
+indent_style = space
+indent_size = 2
 
 [*.{fs,fsi}]
-fsharp_bar_before_discriminated_union_declaration=true
-fsharp_space_before_uppercase_invocation=true
-fsharp_space_before_class_constructor=true
-fsharp_space_before_member=true
-fsharp_space_before_colon=true
-fsharp_space_before_semicolon=true
-fsharp_multiline_bracket_style=aligned
-fsharp_newline_between_type_definition_and_members=true
-fsharp_align_function_signature_to_indentation=true
-fsharp_alternative_long_member_definitions=true
-fsharp_multi_line_lambda_closing_newline=true
-fsharp_experimental_keep_indent_in_branch=true
-fsharp_max_value_binding_width=80
-fsharp_max_record_width=0
-max_line_length=120
-end_of_line=lf
+fsharp_bar_before_discriminated_union_declaration = true
+fsharp_space_before_uppercase_invocation = true
+fsharp_space_before_class_constructor = true
+fsharp_space_before_member = true
+fsharp_space_before_colon = true
+fsharp_space_before_semicolon = true
+fsharp_multiline_bracket_style = aligned
+fsharp_newline_between_type_definition_and_members = true
+fsharp_align_function_signature_to_indentation = true
+fsharp_alternative_long_member_definitions = true
+fsharp_multi_line_lambda_closing_newline = true
+fsharp_experimental_keep_indent_in_branch = true
+fsharp_max_value_binding_width = 80
+fsharp_max_record_width = 0
+max_line_length = 120
+end_of_line = lf
 
 [*.{appxmanifest,build,dtd,nuspec,xaml,xamlx,xoml,xsd}]
-indent_style=space
-indent_size=2
-tab_width=2
+indent_style = space
+indent_size = 2
+tab_width = 2
